Tag failing tests on mobile config as desktop-only
