+++ tests/integration/tests/autocomplete.js
@@ -28,23 +28,6 @@ beforeEach(async () => {
   responseHandler.addPreparedResponse(mockPoi, /places\/*/);
 });
 
-test('search and clear', async () => {
-  responseHandler.addPreparedResponse(mockAutocomplete, /autocomplete\?q=Hello/);
-  await page.goto(APP_URL);
-  await autocompleteHelper.typeAndWait('Hello');
-  expect(await exists(page, '#clear_button_desktop')).toBeTruthy();
-
-  const autocompleteItems = await autocompleteHelper.getSuggestList();
-  expect(autocompleteItems.length).toEqual(SUGGEST_MAX_ITEMS);
-
-  const searchValue = await autocompleteHelper.getSearchInputValue();
-  expect(searchValue).toEqual('Hello');
-
-  await page.click('#clear_button_desktop');
-  const searchValueAfterClear = await autocompleteHelper.getSearchInputValue();
-  expect(searchValueAfterClear).toEqual('');
-});
-
 test('search with no suggest', async () => {
   responseHandler.addPreparedResponse(mockAutocompleteEmpty, /autocomplete\?q=Goodbye/);
   await page.goto(APP_URL);
@@ -214,36 +197,6 @@ test('favorite search', async () => {
   await page.keyboard.type('Hello');
   expect(await exists(page, '.autocomplete_separator_label')).toBeTruthy();
 });
-
-
-// http://idunn_test.test/v1/pois/osm:node:4872758213?lang=fr
-test('submit key', async () => {
-  responseHandler.addPreparedResponse(mockAutocomplete, /autocomplete\?q=Hello/);
-  await page.goto(APP_URL);
-  /* submit with data already loaded */
-  await autocompleteHelper.typeAndWait('Hello');
-  await page.keyboard.press('Enter');
-  await page.waitForSelector('.autocomplete_suggestions', { hidden: true });
-
-  const { center } = await getMapView(page);
-
-  let firstFeatureCenter = mockAutocomplete.features[0].geometry.coordinates;
-  expect(center).toEqual({ lat: firstFeatureCenter[1], lng: firstFeatureCenter[0] });
-  await page.click('#clear_button_desktop');
-
-  /* force specific query */
-  responseHandler.addPreparedResponse(mockAutocompleteAllTypes, /autocomplete\?q=paris/);
-  await autocompleteHelper.typeAndWait('paris');
-  await page.keyboard.press('Enter');
-  await page.waitForSelector('.autocomplete_suggestions', { hidden: true });
-
-  await page.waitFor(300);
-
-  const { center: newCenter } = await getMapView(page);
-  firstFeatureCenter = mockAutocompleteAllTypes.features[0].geometry.coordinates;
-  expect(newCenter).toEqual({ lat: firstFeatureCenter[1], lng: firstFeatureCenter[0] });
-});
-
 
 test('suggestions should not reappear after fast submit', async () => {
   responseHandler.addPreparedResponse(
